requirements.txt
fastapi[all]
uvicorn
pillow
boto3
opencv-python-headless
opencv-contrib-python
